[Update] - CleanUp - Refactored form into different classes for cleaner, more responsive code.
