+++ VisualSQLPro/VisualSQLPro/PopupForm.cs
@@ -4,7 +4,7 @@ using System.Net;
 using System.Net.Sockets;
 using System.Windows.Forms;
 
-namespace VisualSQL
+namespace VisualSQLPro
 {
     public partial class PopupForm: Form
     {
