+++ VisualSQLPro/VisualSQLPro/VisualSQLPro.csproj
@@ -62,6 +62,9 @@
     <Compile Include="ClientForm.Designer.cs">
       <DependentUpon>ClientForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="MetadataLogic.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="PopupForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -70,6 +73,27 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ResizeEvents.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ResizeLogic.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ResizeTimers.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ScintillaSetUp.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="TcpLogic.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ThreadSafety.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="ToolStripMenuLogic.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <EmbeddedResource Include="ClientForm.resx">
       <DependentUpon>ClientForm.cs</DependentUpon>
     </EmbeddedResource>
